+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## XX.XX.XX
+* ! Minor breaking change ! If a manual session is already started, it will not be possible to call "BeginSession" without stopping the previous one 
+* ! Minor breaking change ! If a manual session has not been started, it will not be possible to call "UpdateSession"
+* ! Minor breaking change ! If a manual session has not been started, it will not be possible to call "EndSession"
+
 * Added SDK internal limits configuration interface
 
 * Deprecated the following calls from "CountlyConfig":